requirements.txt
streamlit
spacy
transformers
nltk
PyPDF2
python-docx
matplotlib
seaborn
plotly
pandas
numpy
scikit-learn
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
PyQt6
pdfplumber
python-magic